+++ src/test.html
@@ -6,7 +6,7 @@
     <title>
       Memorandum of understanding and practice journey when implementing
     </title>
-    <script type="module" src="./temple/app.js"></script>
+    <script type="module" src="./rt_temple/app.js"></script>
   </head>
   <body>
     <canvas id="canvas"></canvas>
